requirements.txt
asn1crypto==0.24.0
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
cryptography==2.3.1
Django==1.11
djangorestframework==3.9.0
idna==2.7
ndg-httpsclient==0.5.1
pkg-resources==0.0.0
pusher==2.0.1
pyasn1==0.4.4
pycparser==2.19
pyOpenSSL==18.0.0
six==1.11.0
africastalking==1.1.4
astroid==2.0.4
chardet==3.0.4
confusable-homoglyphs==3.2.0
dj-database-url==0.5.0
django-bootstrap4==0.0.7
django-decouple==2.1
django-heroku==0.3.1
django-registration==2.4.1
django-secretballot==1.0.0
djangorestframework==3.9.0
gunicorn==19.9.0
idna==2.7
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
Pillow==5.3.0
pkg-resources==0.0.0
psycopg2==2.7.5
psycopg2-binary==2.7.5
pylint==2.1.1
python-dateutil==2.7.3
python-decouple==3.1
pytz==2018.5
pyuploadcare==2.5.0
requests==2.19.1
schema==0.6.8
six==1.11.0
typed-ast==1.1.0
urllib3==1.23
whitenoise==4.1
wrapt==1.10.11